fixed test suite
