+++ package.json
@@ -42,5 +42,9 @@
     "supertest": "^5.0.0",
     "webpack": "^4.44.2",
     "webpack-cli": "^3.3.12"
+  },
+  "jest": {
+    "verbose": true,
+    "testEnvironment": "node"
   }
 }
